+++ CHANGELOG.md
@@ -11,5 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump monero-rs to 0.16.0 ([21](https://github.com/monero-ecosystem/monero-rpc-rs/pull/21))
 - Add tests, fix various bugs uncovered by testing
+- Add close wallet method
 
 [Unreleased]: https://github.com/monero-ecosystem/monero-rpc-rs/compare/363c433023318877e9d397dbe2b50bdf88cdee9d...HEAD